==> code/10_MSFraggerOutput_Aggregation.Rmd <==
---
title: "Aggregation of MSFragger output including peptide intensities"
author: "Christoph N. Schlaffner"
output:
  pdf_document: default
  html_document: default
---
```{r, include=FALSE}
options(stringsAsFactors = FALSE)

infile <- "MSFragger_INSOL.csv"
outfile <- "MSFragger_INSOL_agg.csv"
#infile <- "MSFragger_SOL.csv"
#outfile <- "MSFragger_SOL_agg.csv"

indata <- read.csv(infile,header=TRUE)

indata.CTRL <- indata[grep("CTRL",indata$Sample),]
indata <- indata[-grep("CTRL",indata$Sample),]

indata.CTRLagg <- aggregate(indata.CTRL$Intensity,by=list(indata.CTRL$Sample,indata.CTRL$Accessions,indata.CTRL$Sequence,indata.CTRL$ModifiedSequence),FUN="mean")
colnames(indata.CTRLagg) <- c("Sample","Accessions","Sequence","ModifiedSequence","Intensity")


aa <- indata.CTRL[!duplicated(indata.CTRL[,c(1,2,4,5,6,7,8,9,10)]),]
aa$Theor.z <- rep(NA,dim(aa)[1])
aa$Intensity <- NULL

aa <- merge(aa,indata.CTRLagg,all=TRUE)
indata.CTRL <- aa[,colnames(indata.CTRL)]

indata <- rbind(indata,indata.CTRL)

indata.agg <- aggregate(indata$Intensity,by=list(indata$Sample,indata$Accessions,indata$Sequence,indata$ModifiedSequence),FUN="mean")
colnames(indata.agg) <- c("Sample","Accessions","Sequence","ModifiedSequence","Intensity")

aa <- indata[!duplicated(indata[,c(1,2,4,5,6,7,8,9,10)]),]
aa$Theor.z <- rep(NA,dim(aa)[1])
aa$Intensity <- NULL

aa <- merge(aa,indata.agg,all=TRUE)
indata <- aa[,colnames(indata)]

write.csv(indata,file=outfile,quote=FALSE,na="",row.names=FALSE)
```
